+++ app/code/community/Bitbull/Tooso/Block/TrackingPixel.php
@@ -6,7 +6,9 @@
 
 class Bitbull_Tooso_Block_TrackingPixel extends Mage_Core_Block_Template
 {
-    const SCRIPT_ENDPOINT = '/tooso/tracking/pixel/';
+    const BLOCK_ID = 'tooso_tracking_pixel';
+    const SCRIPT_ID = 'tooso-tracking-script';
+    const SCRIPT_ENDPOINT = 'tooso/tracking/';
 
     /**
      * @var Bitbull_Tooso_Helper_Log
@@ -14,9 +16,14 @@ class Bitbull_Tooso_Block_TrackingPixel extends Mage_Core_Block_Template
     protected $_logger = null;
 
     /**
-     * @var null|integer
+     * @var null|string
      */
-    protected $_product_id = null;
+    protected $_currentPage = null;
+
+    /**
+     * @var null|string
+     */
+    protected $_lastPage = null;
 
     public function _construct()
     {
@@ -24,25 +31,61 @@ class Bitbull_Tooso_Block_TrackingPixel extends Mage_Core_Block_Template
 
         $this->_logger = Mage::helper('tooso/log');
 
-        $this->setBlockId('tooso_tracking_pixel');
+        $this->setBlockId(self::BLOCK_ID);
         $this->addCacheTag(array(
             Mage::app()->getStore()->getId(),
             Mage_Catalog_Model_Product::CACHE_TAG
+        ));
+        $this->addData(array(
+            'cache_lifetime' => null,
         ));
     }
 
     protected function _toHtml()
     {
-        if($this->_product_id == null){
-            $this->_logger->warn('Tracking script: product_id not set');
-            return;
-        }
-        $url = self::SCRIPT_ENDPOINT."product/".$this->_product_id;
-        $this->_logger->debug('Tracking script: added tracking script');
-        return "<script async type='text/javascript' src='".$url."'></script>";
+        return;
     }
 
-    public function setProductID($id){
-        $this->_product_id = $id;
+    /**
+     * Get cache identifier
+     *
+     * @return string
+     */
+    protected function _getCacheId()
+    {
+        return strtoupper(self::BLOCK_ID);
+    }
+
+
+    /**
+     * @param $url string
+     */
+    public function setCurrentPage($url){
+        $this->_currentPage = $url;
+    }
+
+    /**
+     * @param $url string
+     */
+    public function setLastPage($url){
+        $this->_lastPage = $url;
+    }
+
+    /**
+     * @return string
+     */
+    protected function _getPageParams(){
+        $url = "currentPage/".base64_encode($this->_currentPage);
+        if($this->_lastPage != null){
+            $url .= "/lastPage/".base64_encode($this->_lastPage);
+        }
+        return $url;
+    }
+
+    /**
+     * @param $id
+     */
+    public function setObjectID($id){
+
     }
 }