+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0]
+- Added API v2 compatibility
+- Catch current_layer null error
+- Substituted "+" with " " in query string
+- Soft error when 'before_body_end' block not exist
+- Refactory search id using cookies
+
+## [2.0.1]
+- Get base url to point correct controller endpoint
+
 ## [2.0.0]
 - Disable email report on empty query
 - Increase API response timeout